rename matchWithPartners file
